python/requirements.txt
gpiod==2.4.0
gpiodevice==0.0.5
inky==2.2.1
numpy==2.3.4
pillow==12.0.0
smbus2==0.5.0
spidev==3.8
